Add {Risk,Protocol,Market}Parameter, Local, Order, and Position lib tests (#26)
* Add 'Local' lib tests

* Add 'MarketParameter' lib tests

* Add 'ProtocolParameter' lib tests

* Add 'RiskParameter' lib tests

* Add 'Position' lib tests

* Add 'Order' lib tests

* [RiskParameterTest] Use different values for structs

+++ packages/perennial/contracts/types/RiskParameter.sol
@@ -140,7 +140,7 @@ library RiskParameterStorageLib {
         if (newValue.utilizationCurve.minRate.gt(UFixed6.wrap(type(uint32).max))) revert RiskParameterStorageInvalidError();
         if (newValue.utilizationCurve.maxRate.gt(UFixed6.wrap(type(uint32).max))) revert RiskParameterStorageInvalidError();
         if (newValue.utilizationCurve.targetRate.gt(UFixed6.wrap(type(uint32).max))) revert RiskParameterStorageInvalidError();
-        if (newValue.utilizationCurve.targetUtilization.gt(UFixed6.wrap(type(uint32).max))) revert RiskParameterStorageInvalidError();
+        if (newValue.utilizationCurve.targetUtilization.gt(UFixed6.wrap(type(uint24).max))) revert RiskParameterStorageInvalidError();
         if (newValue.pController.k.gt(UFixed6.wrap(type(uint40).max))) revert RiskParameterStorageInvalidError();
         if (newValue.pController.max.gt(UFixed6.wrap(type(uint32).max))) revert RiskParameterStorageInvalidError();
         if (newValue.minMaintenance.gt(UFixed6.wrap(type(uint48).max))) revert RiskParameterStorageInvalidError();
